requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
csv-reader==1.2.0
idna==3.7
joblib==1.4.0
numpy==1.26.4
pandas==2.2.2
pip==22.0.2
pyTelegramBotAPI==4.17.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
scikit-learn==1.4.2
scipy==1.13.0
setuptools==59.6.0
six==1.16.0
threadpoolctl==3.4.0
tzdata==2024.1
urllib3==2.2.1
wheel==0.43.0
